Assignment 6
I did a lot in this assignment, I actually fixed the file directory and how it works (i'll figure out how to do it with the other pull requests in the future) I made test cases for my secondary and my kernel, and I also made two demos so it can be seen how to use this. This was mainly made for personal and fun usage but it can also be used in an academic setting such as an Anime club or another club that wants to keep track of shows or even manga.